requirements.txt
openai==1.101.0
pymupdf==1.26.4
PyPDF2==3.0.1
python-dotenv==1.1.0
tiktoken==0.11.0
pyyaml==6.0.2
anthropic>=0.40.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
streamlit>=1.35.0
